refactor(rob-269-p3): lift us_action_report → action_report/us (clean-cut)
Phase 3 of ROB-269 — pre-plan Decision 2 lift. Moves the 6 files of
app/services/us_action_report/ to app/services/action_report/us/ with
no transitional re-export shim (memory: clean-cut preference).

* Source: ``git mv app/services/us_action_report app/services/action_report/us``
* Internal imports inside us/__init__.py updated to point at the new
  package path.
* Test imports updated (5 files); test file locations kept under
  tests/test_us_action_report_*.py to minimise diff churn — a follow-up
  can relocate to tests/services/action_report/us/ if desired.

No call sites in routers/MCP/services referenced us_action_report
outside of its own __init__, so the lift is contained. 22 lifted tests
still pass without behaviour change.
