+++ app/services/action_report/us/new_buy_candidates.py
