+++ tests/test_us_action_report_action_classifier.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import UTC, datetime, timedelta
 
 from app.schemas.us_action_report import KISUSAccountSnapshot, USHolding, USOpenOrder
-from app.services.us_action_report.action_classifier import (
+from app.services.action_report.us.action_classifier import (
     build_us_held_position_action_cards,
 )
