+++ tests/test_us_action_report_discord_formatter.py
@@ -8,13 +8,13 @@ from app.schemas.us_action_report import (
     USHolding,
     USOpenOrder,
 )
-from app.services.us_action_report.action_classifier import (
+from app.services.action_report.us.action_classifier import (
     build_us_held_position_action_cards,
 )
-from app.services.us_action_report.discord_formatter import (
+from app.services.action_report.us.discord_formatter import (
     build_us_action_report_discord_message,
 )
-from app.services.us_action_report.new_buy_candidates import (
+from app.services.action_report.us.new_buy_candidates import (
     build_us_new_buy_candidate_cards,
 )
