+++ tests/test_us_action_report_order_preview.py
@@ -13,8 +13,8 @@ from app.schemas.us_action_report import (
     USHolding,
     USOpenOrder,
 )
-from app.services.us_action_report import order_preview
-from app.services.us_action_report.order_preview import (
+from app.services.action_report.us import order_preview
+from app.services.action_report.us.order_preview import (
     preview_kis_us_live_order,
     submit_kis_us_live_order_from_preview_disabled,
 )
